// src/app/openapi/model/gomdonLogs.ts
/**
 * Gom Don API
 * Gom Don API.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: huynhnhon.dev@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GomdonLogs { 
    content: string;
    id: number;
    name: string;
    type: number;
    uid: string;
}